Debianize Kafka
In order to debianize kafka sensibly I had to create my own build system
based on Makefiles. These are stored in debian/patches source directory.
The key issue here is that per analytics request the branch chosen to
develop against was trunk since we will be using version 0.8 which has
not yet been released. The package depends on default-jre, adduser,
scala library and on all jar files that can be used from the system.
One point of concern is the velocity package which depends on ant, a
package we probably don't want installed in production.
Note: package libjoptsimple-java from precise has not been used because
the package would not compile with version 3.1 but only with 3.2

Change-Id: I55e65084eb5444a03e63cc479f9f3b72cf07a146